app/backend/requirements.txt
fastapi==0.109.0
pydantic==2.5.3
uvicorn==0.15.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0